Arduino code added.
Arduino code will listen for data on the serial communication line and enable/disable security mode for the user, depending on the user's selection on the Android app. Once enabled, the on-board system will utilize a microwave sensor to detect nearby moving objects, such as a human approaching the device. The system will check if there is an approaching object every 10 seconds. If an object is detected, the arduino will send a message over Bluetooth to the user's Android phone, notifying them of a tripped alarm.